lambda/requirements.txt
numpy
pillow
tensorflow==2.9.1
